+++ python_modules/dagster/dagster/config/post_process.py
@@ -19,6 +19,15 @@ def post_process_config(config_type, config_value):
     return _recursively_process_config(ctx, config_value)
 
 
+def resolve_defaults(config_type, config_value):
+    ctx = ValidationContext(
+        config_type=check.inst_param(config_type, 'config_type', ConfigType),
+        stack=EvaluationStack(config_type=config_type, entries=[]),
+    )
+
+    return _recursively_resolve_defaults(ctx, config_value)
+
+
 def _recursively_process_config(context, config_value):
     evr = _recursively_resolve_defaults(context, config_value)