+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 **Breaking Changes**
 
+- `default_value` in `Field` no longer accepts native instances of python enums. Instead
+  the underlying string representation in the config system must be used.
 - `default_value` in `Field` no longer accepts callables.
 - The `dagster_aws` imports have been reorganized; you should now import resources from
   `dagster_aws.<AWS service name>`. `dagster_aws` provides `s3`, `emr`, `redshift`, and `cloudwatch`